[Data] Address review: move pure guard tests to unit/, doc C++ accessors
- Move the pure-logic OutputBackpressureGuard tests (task ordering, bypass
  lane) into tests/unit/test_output_backpressure_guard.py, where the unit
  conftest enforces no Ray runtime / sleep. The two tests that drive
  process_completed_tasks (and need ray.wait) stay in the integration file.
- Hoist the shared FakeDataOpTask fake and the mock-op factory into
  tests/util.py so both the unit and integration tests reuse one helper.
- Add Doxygen \return tags to the new CoreWorker / task-submitter accessors,
  matching the neighboring QueueGeneratorForResubmit declarations.
