+++ python/ray/data/tests/test_streaming_executor.py
@@ -76,7 +76,11 @@ from ray.data.block import BlockAccessor, BlockMetadataWithSchema, TaskExecWorke
 from ray.data.context import EXECUTION_CALLBACKS_ENV_VAR, DataContext
 from ray.data.tests.conftest import *  # noqa
 from ray.data.tests.conftest import noop_counter
-from ray.data.tests.util import fetcher_has_pending_work
+from ray.data.tests.util import (
+    FakeDataOpTask,
+    fetcher_has_pending_work,
+    make_backpressured_op,
+)
 from ray.exceptions import GetTimeoutError
 
 
@@ -358,27 +362,6 @@ def test_process_completed_tasks_inline(sleep_task_ref, ray_start_regular_shared
     o2.mark_execution_finished.assert_called_once()
 
 
-class _FakeDataOpTask(DataOpTask):
-    """Minimal ``DataOpTask`` stand-in recording the byte budget of every read."""
-
-    def __init__(self, task_index, task_id, bytes_read, waitable=None, calls=None):
-        self._task_index = task_index
-        self._task_id = task_id
-        self._bytes_read = bytes_read
-        self._waitable = waitable
-        self.calls = calls if calls is not None else []
-
-    def get_waitable(self):
-        return self._waitable
-
-    def get_task_id(self):
-        return self._task_id
-
-    def on_data_ready(self, max_bytes_to_read, metadata_fetcher):
-        self.calls.append((self._task_id, max_bytes_to_read))
-        return self._bytes_read
-
-
 class _ZeroOutputBudgetPolicy:
     """Simulate output backpressure preventing ordinary task-output reads."""
 
@@ -390,21 +373,6 @@ class _ZeroOutputBudgetPolicy:
         return 0
 
 
-def _make_op(bypass_blocks: int) -> MagicMock:
-    op = MagicMock()
-    op.data_context.lineage_reconstruction_backpressure_bypass_blocks = bypass_blocks
-    return op
-
-
-def _guard() -> OutputBackpressureGuard:
-    """A guard for exercising its stateless task-ordering / bypass-lane policy.
-
-    ``order_ready_tasks`` and ``reconstruction_bypass_lane`` take the queued
-    generator set as an argument, so no per-round setup is needed.
-    """
-    return OutputBackpressureGuard({}, MagicMock())
-
-
 def test_process_completed_tasks_skips_core_worker_query_when_nothing_is_output_limited(
     ray_start_regular_shared,
 ):
@@ -412,7 +380,7 @@ def test_process_completed_tasks_skips_core_worker_query_when_nothing_is_output_
     # only issued when some operator has a finite output read budget. With no
     # limiting policy there is no budget, so even with a ready task the query
     # must be skipped.
-    task = _FakeDataOpTask(0, "normal", bytes_read=0, waitable=ray.put("normal"))
+    task = FakeDataOpTask(0, "normal", bytes_read=0, waitable=ray.put("normal"))
     op = MagicMock()
     op.get_active_tasks.return_value = [task]
     op.has_next.return_value = False
@@ -435,100 +403,15 @@ def test_process_completed_tasks_skips_core_worker_query_when_nothing_is_output_
     assert task.calls == [("normal", None)]
 
 
-def test_order_ready_tasks_puts_queued_generators_first():
-    gen = _FakeDataOpTask(2, "gen", bytes_read=10)
-    normal = _FakeDataOpTask(0, "normal", bytes_read=10)
-    meta = MagicMock(spec=MetadataOpTask)
-    meta.task_index.return_value = 1
-
-    # The queued-for-resubmit generator claims the operator's output budget
-    # first, even though its task index is highest. Non-data tasks (no task ID)
-    # are ordered without being probed for eligibility.
-    assert _guard().order_ready_tasks([normal, meta, gen], {"gen"}) == [
-        gen,
-        normal,
-        meta,
-    ]
-
-    # With nothing queued, ordering falls back to plain task-index order.
-    assert _guard().order_ready_tasks([gen, meta, normal], set()) == [
-        normal,
-        meta,
-        gen,
-    ]
-
-
-@pytest.mark.parametrize("op_output_budget", [0, 1])
-def test_reconstruction_bypass_lane_granted_when_fully_backpressured(op_output_budget):
-    # Budget 0 is full backpressure; budget 1 is what ``should_unblock`` bumps it
-    # to, which is still too little for an ordinary task to make progress.
-    gen = _FakeDataOpTask(0, "gen", bytes_read=10)
-    lane = _guard().reconstruction_bypass_lane(_make_op(3), op_output_budget, {"gen"})
-    assert lane.try_drain(gen, metadata_fetcher=None)
-    # Drained one block at a time so the bypass stays bounded.
-    assert gen.calls == [("gen", 1)] * 3
-
-
-@pytest.mark.parametrize("op_output_budget", [None, 2, 1024])
-def test_no_reconstruction_bypass_lane_when_output_budget_allows_progress(
-    op_output_budget,
-):
-    gen = _FakeDataOpTask(0, "gen", bytes_read=10)
-    lane = _guard().reconstruction_bypass_lane(_make_op(3), op_output_budget, {"gen"})
-    assert not lane.try_drain(gen, metadata_fetcher=None)
-    assert gen.calls == []
-
-
-def test_reconstruction_bypass_lane_disabled_by_zero_bypass_blocks():
-    gen = _FakeDataOpTask(0, "gen", bytes_read=10)
-    lane = _guard().reconstruction_bypass_lane(_make_op(0), 0, {"gen"})
-    assert not lane.try_drain(gen, metadata_fetcher=None)
-    assert gen.calls == []
-
-
-def test_reconstruction_bypass_lane_ignores_ordinary_tasks():
-    normal = _FakeDataOpTask(0, "normal", bytes_read=10)
-    lane = _guard().reconstruction_bypass_lane(_make_op(3), 0, {"gen"})
-    assert not lane.try_drain(normal, metadata_fetcher=None)
-    assert normal.calls == []
-
-
-def test_reconstruction_bypass_lane_stops_when_generator_has_no_ready_output():
-    # The lane pulls one block at a time up to the configured cap, but stops as
-    # soon as ``on_data_ready`` reports 0 bytes read (generator drained / pending
-    # emits / nothing buffered), so it never spins the full cap for nothing.
-    gen = _FakeDataOpTask(0, "gen", bytes_read=0)
-    lane = _guard().reconstruction_bypass_lane(_make_op(100), 0, {"gen"})
-    assert lane.try_drain(gen, metadata_fetcher=None)
-    assert gen.calls == [("gen", 1)]
-
-
-def test_reconstruction_bypass_lane_budget_is_shared_per_operator():
-    # ``lineage_reconstruction_backpressure_bypass_blocks`` is a per-operator
-    # budget shared across that operator's queued-for-resubmit generators, NOT
-    # granted N-per-task: with two queued generators and a cap of 3, the total
-    # is 3 bypass reads, not 6.
-    gen_a = _FakeDataOpTask(0, "gen_a", bytes_read=10)
-    gen_b = _FakeDataOpTask(1, "gen_b", bytes_read=10)
-    lane = _guard().reconstruction_bypass_lane(_make_op(3), 0, {"gen_a", "gen_b"})
-
-    assert lane.try_drain(gen_a, metadata_fetcher=None)
-    assert gen_a.calls == [("gen_a", 1)] * 3
-    # gen_b finds the shared allowance exhausted and falls back to the
-    # operator's (zeroed) byte budget.
-    assert not lane.try_drain(gen_b, metadata_fetcher=None)
-    assert gen_b.calls == []
-
-
 @pytest.mark.parametrize("bypass_blocks", [1, 3])
 def test_process_completed_tasks_drains_queued_generator_resubmit_tasks_under_backpressure(
     ray_start_regular_shared, bypass_blocks
 ):
     calls = []
-    normal_task = _FakeDataOpTask(
+    normal_task = FakeDataOpTask(
         0, "normal", bytes_read=0, waitable=ray.put("normal"), calls=calls
     )
-    queued_resubmit_task = _FakeDataOpTask(
+    queued_resubmit_task = FakeDataOpTask(
         1,
         "queued_resubmit",
         bytes_read=10,
@@ -536,7 +419,7 @@ def test_process_completed_tasks_drains_queued_generator_resubmit_tasks_under_ba
         calls=calls,
     )
 
-    op = _make_op(bypass_blocks)
+    op = make_backpressured_op(bypass_blocks)
     op.get_active_tasks.return_value = [normal_task, queued_resubmit_task]
     op.has_next.return_value = False
