+++ src/ray/core_worker/task_submission/normal_task_submitter.h
@@ -150,6 +150,7 @@ class NormalTaskSubmitter {
 
   /// Return running streaming generator tasks that have a queued resubmission for
   /// lineage reconstruction once their current attempt finishes.
+  /// \return The task IDs of those streaming generators.
   std::vector<TaskID> GetQueuedGeneratorResubmitTaskIds() const;
 
   /// Check that the scheduling_key_entries_ hashmap is empty by calling the private
